Server only have to decipher base 64 encoded strings
Once deciphered they have to find the plaintext inside the gibbrish.
It is marked by '$' before and after.
Also added a check to see if there is a "end_url+" at the beginning
of the PT, if so they get the data and then rerun a "get_address" to
know where to send the data